refactor stack into iterator, part 1
we can eliminate allocations by treating the entire stack as an iterator
over results, and collecting it into a result at the end. this process
is a tad awkward due to interactions between ? and the various iterator
combinators, but somehow i actually find this just as legible as the
imperative version. maybe i'm crazy?