End "fixup!" commit messages with newline
This seems to be the standard behavior of other git tools.

In particular, `git commit -m "..."` adds the newline, which in fact
motivated this change - this way I can assert that the commit created
by git-absorb is identical to the manually crafted commit in my manual
testing that involves running git commands from a script.