Use Rust 2018 edition
No particular reason for the switch, I tried it when I got a bunch of
errors related to bitflags when upgrading git2 (see the following
commits). Turned out they were not related to the edition used, but
there's no reason not to switch anyway, is there?

This requires Rust 1.31 (from last year) to build.