Fix --and-rebase when last commit is the repository root
When the last commit in the stack is the repository root, this assertion
fails because it expects the last commit to always have exactly one
parent. This patch fixes the assertion to allow a base commit with zero
parents (i.e., the repository root) and in that case executes `git
rebase` with the `--root` option so we can absorb changes into the root
commit.