bump git2/libgit2-sys locked dependencies
There is a double free in libgit2 that is fixed somewhere between
v1.0.0 and v1.1.0. This means that the transitive libgit2-sys
dependency needs to be after 0.12.14+1.1.0.

This manifests itself via the following error message on certain
repositories:

    $ git absorb
    free(): double free detected in tcache 2
    Aborted (core dumped)

Bump git2 to the latest version, which also updates the transitive
libgit2-sys (and libgit2) dependency.

Fixes #69