src: lib.rs: add whole-file match option
Sometimes we just forget some change (eg: intializing some object) that
happens in a completely unrelated place with the current hunks. In those
cases, might be helpful to just match the first commit touching the same
file as the current hunk.

This above is enabled by passing the new '-w|--whole-file' flag. It
should be used with extreme care to not get undesired results.

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>