Add option to create one fixup per commit.
This means that all hunks meant to be fixed up into the same commit
will be grouped into a single commit, rather than multiple ones.

In order to not break the workflow of others, we choose to make this
an option rather than the default. Moreover, folks can `alias
git-absorb="git-absorb --one-fixup-per-commit"` if they wish to always
use this option, and not have to type it out.