autostage: automatically add everything to index if index was empty
often times, i just want to try and absorb everything.

right now, this means an extra step of `git add .` before `git absorb`,
and potentially a `git reset @` later if not everything was fixed up.

in this patch we remove both of the extra step for convenience:
if nothing was staged, we stage everything, try to create fixups,
and whatever's left uncommitted - we unstage.