Don't sort hunks based on dest_commit.
Because each hunk needs to be displaced based on the previously
applied hunk, sorting hunks will change the order in which they are
applied, therefore undoing all our previous calculations.

This fixes #116, but breaks the --one-fixup-per-commit feature in
cases where there is an odd hunk in between some hunks that could've
been merged together.