Support annotated tags given as base ref
Improve the types of refspecs git-absorb knows how to parse that are
given via the '--base' flag.

This fixes the following:

    # create an annotated tag
    $ git tag -a 'annotated-tag' -m 'My annotated tag description'

    # make + stage some changes
    (...)

    # try to absorb the changes
    $ git absorb --base annotated-tag

According to the Git docs:

    > Git supports two types of tags: lightweight and annotated.
    >
    > A lightweight tag is very much like a branch that doesn't
    > change-it's just a pointer to a specific commit.
    >
    > Annotated tags, however, are stored as full objects in the Git
    > database.

This meant that when passing an annotated tag via '--base', calling
git2::Repository::find_commit would fail since the given object would
be of the incorrect type:

    > the requested type does not match the type in the ODB

However, eventually if we keep unpeeling, we would encounter the
corresponding commit.

Use git2::Object::peel_to_commit instead of simply peeling once
via 'find_commit' to ensure that peeling is performed recursively
until a commit is found.