Remove slog-async dependency
According to the slog-term docs:

    > PlainSync Decorator implementation
    >
    > This implementation is exactly like PlainDecorator but it takes
    > care of synchronizing writes to io.

Since we care about colouring, wrap a slog_term::TermDecorator in a
std::sync::Mutex so that coloured output is still maintained.

This ensures that log messages will not be dropped when the logging
rate exceeds the backend's ability to drain (as the previous
slog-async backend could do). Previously when this occurred, the log
output would show the following:

    slog-async: logger dropped messages due to channel overflow

An alternative might be to keep the slog-async dependency but use
slog_async::OverflowStrategy::Block to ensure that overflowing writes
will block, rather than forcing each write through the mutex.

Fixes #131