Warn about non-modified patches
Some patches do not modify an existing file in place.
Instead they add, remove, or rename a file.
This is a "non-modified patch", and will commute with
everything in the stack.

If all patches are non-modified, then no changes will be absorbed, so
warn the user and stop processing. Even users that auto-stage changes
will be warned, since they would've run git-absorb with the expectation
that it would do something.

Otherwise, if only some patches are non-modified, we can continue
processing with a warning. In this case, do not warn users who
auto-stage changes, since they may routinely keep untracked files in
their working directory.

Before this,

1. users who did not auto-stage were warned when nothing was staged.
   We have a more explicit warning and early exit above now.
2. there were staged changes (auto-staged or not) and all the patches
   were non-modified, we'd warn "Could not find a commit to fix up",
   which is accurate, but not as informative.